+++ src/pages/Trade.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import TradeWidget from 'components/TradeWidget'
-import { PageWrapper } from 'components/layout/PageWrapper'
+import { PageWrapper } from 'components/layout/PageWrappers'
 
 const Trade: React.FC = () => (
   <PageWrapper>
